refactor: add new object in new file 'InputView.kt' to extract 'readCarNames()'
Extract 'readCarNames()' function from Application.kt file to new InputView.kt file and wrapped in new object InputView.
